feature_extractor/torch_stft/requirements.txt
torch==1.13.1
onnx-simplifier==0.4.36
matplotlib==3.9.2
numpy==1.26.4
soundfile==0.12.1
